chore: unify env utilities
