+++ src/lib/env.ts
@@ -27,4 +27,25 @@ export function requireEnv(names: string[]) {
   }
 }
 
+/**
+ * Resolve repository configuration from environment variables.
+ *
+ * Required:
+ *   - TARGET_OWNER (e.g. "basstian-ai")
+ *   - TARGET_REPO (e.g. "simple-pim-1754492683911")
+ */
+export function parseRepo(): { owner: string; repo: string } {
+  const targetOwner = process.env.TARGET_OWNER;
+  const targetRepo = process.env.TARGET_REPO;
+
+  if (!targetOwner || !targetRepo) {
+    throw new Error("Missing required TARGET_OWNER and TARGET_REPO environment variables");
+  }
+
+  return {
+    owner: targetOwner,
+    repo: targetRepo,
+  };
+}
+
 // Avoid requiring SUPABASE variables for commands that don't need them.
